refactor: Refactor to short code
